feat: move deps file
